+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Run bin/setup after DB creation
 - Adds `db:fake_data:load` task for development data. It runs after `db:setup` on development environment
+- Configure Sentry as Error Reporting tool
+- Configure clockwork as scheduler
 - Configure mailer for development environment
 - Configure Rspec to work with Paperclip
 - Update [multi buildpack]([multi-buildpack-heroku]) buildback to the heroku version